✨ (xml): add new --dirty and --ignore-halves options
The --dirty option should help you to decide which XHTML tags should be removed or replaced and which should be kept. The --ignore-halves allow you to ignore rows with one column empty, so they're not added to previous row. In this was added Cheerio to ensure that all elements are closed in the final XML file. Also </xml> tag was removed from output file as it is not valid.